Sessions/CheckUser-Secured

+++ classes/JokeDB/Controllers/Joke.php
@@ -53,12 +53,20 @@ class Joke {
 
     }
     public function delete () {
+        $author = $this->authentication->getUser();
+        $joke = $this->jokesTable->findById($_POST['id']);
+        if ($joke['authorId'] != $author['id']) return;
+
         $this->jokesTable->delete($_POST['id']);
 
 	    header('location: /joke/list');
     }
     public function saveEdit () {
         $author = $this->authentication->getUser();
+        if (isset($_GET['id'])) {
+            $joke = $this->jokesTable->findById($_GET['id']);
+            if ($joke['authorId'] != $author['id'])  return;
+        }
         $joke = $_POST['joke'];
         $joke['jokedate'] = new \DateTime();
         $joke['authorId'] = $author['id'];
@@ -78,7 +86,8 @@ class Joke {
                 'template' => 'editjoke.html.php', 
                 'title' => $title,
                 'variables' => [
-                    'joke' => $joke ?? null ]
+                    'joke' => $joke ?? null,
+                    'userId' => $author['id'] ?? null ]
             ];
         
     }
